requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
distro==1.9.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.1.3
langchain-openai==1.1.2
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.15
langsmith==0.4.59
matplotlib-inline==0.2.1
openai==2.11.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
wcwidth==0.2.14
xxhash==3.6.0
zstandard==0.25.0
